requirements.txt
sqlalchemy==2.0.21
python-dotenv==1.0.0
psycopg2==2.9.7